fix JWT token on new User
