+++ server/src/schemas/typeDefs.ts
@@ -5,6 +5,7 @@ export const typeDefs = gql`
     _id: ID
     username: String
     email: String
+    token: String
   }
 
   type Auth {
@@ -57,7 +58,7 @@ export const typeDefs = gql`
   }
 
   type Mutation {
-    addUser(username: String!, email: String!, password: String!): User
+    addUser(username: String!, email: String!, password: String!): Auth
     login(email: String!, password: String!): User
     addChoice(_id: ID!, choice_name: String!): Poll
     rankedChoice(user_id: ID!): Choices
